Bewegungs Methoden bei Characker Klasse eingefügt
